InitWizard: move to its own directory
We're going to refactor this so that each page is its own
QWizardPage subclass, so we need some room to work here.
